requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
psycopg2-binary==2.9.9
werkzeug==3.0.1
python-dotenv==1.0.0
openpyxl==3.1.2
mercadopago==2.2.1
pillow==10.2.0
